requirements.txt
redis.asyncio
fastapi
asyncio
pydantic
typing
google-generativeai